+++ scripts/docs_snippets/verbosity_warning.py
@@ -7,4 +7,4 @@ cf = CountedFloat(2.5)
 with FlopCountingContext(verbosity=Verbosity.WARNING):
     for _ in range(1000):
         _ = math.ldexp(cf, 3)
-    _ = math.isclose(cf, 2.5)
+    _ = math.frexp(cf)